Fix bug in tosc_reset
tosc_reset was operating on offsets based on the location of the
format (o->format).  When it came time to adjust the marker
location to a multiple of 4 bytes it adjusted that index to
a multiple of 4 bytes, which gives you a location aligned
relative to o->format.  But the location needs aligned relative
to the entire buffer.

This change re-writes tosc_reset to operate entirely relative to
the buffer start rather than to the start of the format.